De-inline some nonstd impls
