+++ CMakeLists.txt
@@ -263,7 +263,7 @@ else()
   list(APPEND CCHOST_SOURCES src/host/snmalloc.cpp)
 endif()
 
-list(APPEND CCHOST_SOURCES ${CCF_DIR}/src/host/main.cpp)
+list(APPEND CCHOST_SOURCES ${CCF_DIR}/src/host/main.cpp ${CCF_DIR}/src/ds/nonstd.cpp)
 
 if(COMPILE_TARGET STREQUAL "sgx")
   list(APPEND CCHOST_SOURCES ${CCF_GENERATED_DIR}/ccf_u.cpp)
@@ -615,7 +615,7 @@ add_custom_target(ccf ALL)
 
 set(CCF_IMPL_SOURCE
     ${CCF_DIR}/src/enclave/main.cpp ${CCF_DIR}/src/enclave/enclave_time.cpp
-    ${CCF_DIR}/src/enclave/thread_local.cpp ${CCF_DIR}/src/node/quote.cpp
+    ${CCF_DIR}/src/enclave/thread_local.cpp ${CCF_DIR}/src/node/quote.cpp ${CCF_DIR}/src/ds/nonstd.cpp
 )
 
 if(COMPILE_TARGET STREQUAL "sgx")
@@ -1009,6 +1009,7 @@ if(BUILD_TESTS)
 
     add_test_bin(
       kp_cert_test ${CMAKE_CURRENT_SOURCE_DIR}/src/crypto/test/kp_cert.cpp
+      ${CMAKE_CURRENT_SOURCE_DIR}/src/ds/nonstd.cpp
     )
     target_link_libraries(kp_cert_test PRIVATE ${CMAKE_THREAD_LIBS_INIT})
 
@@ -1082,7 +1083,9 @@ if(BUILD_TESTS)
     )
 
     # Merkle Tree memory test
-    add_executable(merkle_mem src/node/test/merkle_mem.cpp)
+    add_executable(merkle_mem src/node/test/merkle_mem.cpp
+    src/ds/nonstd.cpp
+    )
     target_compile_options(merkle_mem PRIVATE ${COMPILE_LIBCXX})
     target_link_libraries(
       merkle_mem PRIVATE ${CMAKE_THREAD_LIBS_INIT} ${LINK_LIBCXX}
@@ -1093,6 +1096,8 @@ if(BUILD_TESTS)
     add_executable(
       raft_driver ${CMAKE_CURRENT_SOURCE_DIR}/src/consensus/aft/test/driver.cpp
                   src/enclave/thread_local.cpp
+    src/ds/nonstd.cpp
+
     )
     target_link_libraries(raft_driver PRIVATE ccfcrypto.host)
     target_include_directories(raft_driver PRIVATE src/aft)
