+++ cmake/common.cmake
@@ -3,7 +3,9 @@
 
 # Unit test wrapper
 function(add_unit_test name)
-  add_executable(${name} ${CCF_DIR}/src/enclave/thread_local.cpp ${ARGN})
+  add_executable(${name} ${CCF_DIR}/src/enclave/thread_local.cpp 
+    ${CCF_DIR}/src/ds/nonstd.cpp
+  ${ARGN})
   target_compile_options(${name} PRIVATE ${COMPILE_LIBCXX})
   target_include_directories(
     ${name} PRIVATE src ${CCFCRYPTO_INC} ${CCF_DIR}/3rdparty/test
@@ -349,7 +351,9 @@ function(add_picobench name)
     PARSE_ARGV 1 PARSED_ARGS "" "" "SRCS;INCLUDE_DIRS;LINK_LIBS"
   )
 
-  add_executable(${name} ${PARSED_ARGS_SRCS})
+  add_executable(${name} 
+  src/ds/nonstd.cpp
+  ${PARSED_ARGS_SRCS})
 
   target_include_directories(${name} PRIVATE src ${PARSED_ARGS_INCLUDE_DIRS})
