+++ src/http/http2_utils.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/ds/nonstd.h"
 #include "ccf/http_header_map.h"
+#include "ds/nonstd.h"
 
 #include <nghttp2/nghttp2.h>
 #include <optional>
