+++ tests/perf-system/submitter/CMakeLists.txt
@@ -3,6 +3,7 @@ set(SUBMITTER_DIR ${CCF_DIR}/tests/perf-system/submitter)
 add_executable(
   submit ${SUBMITTER_DIR}/submit.cpp ${SUBMITTER_DIR}/handle_arguments.h
          ${SUBMITTER_DIR}/parquet_data.h
+         ${CCF_DIR}/src/ds/nonstd.cpp
 )
 
 add_library(stdcxxhttp_parser.host "${HTTP_PARSER_SOURCES}")
